#1 refactoring and added ensemble types
