+++ src/main/java/com/gigsterous/api/model/Ensemble.java
@@ -6,6 +6,8 @@ import java.util.Set;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
@@ -13,6 +15,9 @@ import javax.persistence.JoinTable;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToMany;
 import javax.persistence.Table;
+
+import com.gigsterous.api.model.enums.EnsembleType;
+import com.gigsterous.api.model.enums.Gender;
 
 @Entity
 @Table(name = "ensembles")
@@ -25,6 +30,10 @@ public class Ensemble {
 
 	@Column(name = "name")
 	private String name;
+	
+	@Column(name = "ensemble_type")
+	@Enumerated(EnumType.STRING)
+	private EnsembleType ensembleType;
 
 	@ManyToMany(cascade = CascadeType.ALL)
 	@JoinTable(name = "users_ensembles", joinColumns = @JoinColumn(name = "ensemble_id", referencedColumnName = "ensemble_id"), inverseJoinColumns = @JoinColumn(name = "user_id", referencedColumnName = "user_id"))
@@ -48,6 +57,14 @@ public class Ensemble {
 
 	public void setName(String name) {
 		this.name = name;
+	}
+
+	public EnsembleType getEnsembleType() {
+		return ensembleType;
+	}
+
+	public void setEnsembleType(EnsembleType type) {
+		this.ensembleType = type;
 	}
 
 	public Set<User> getUsers() {
