+++ src/main/java/com/gigsterous/api/model/enums/EnsembleType.java
@@ -1,0 +1,6 @@
+package com.gigsterous.api.model.enums;
+
+public enum EnsembleType {
+	BAND,
+	ORCHESTRA
+}
